+++ apps/game/server/npc/ai-tasks/wander.ts
@@ -1,6 +1,6 @@
 import { randomItem } from "@mp/std";
 import type { TimeSpan } from "@mp/time";
-import type { Task, TaskInput } from "./Task";
+import type { Task, TaskInput } from "./task";
 
 export function createWanderTask(
   endTime: TimeSpan,
